fix(npm): Remove .npmignore (#3)
* fix(npm): Remove .npmignore

The presence of the .npmignore-file caused npm to ignore the gitignore,
causing it to publish the .env-file which was gitignored but not
npmignored, leaking secrets. Removing this file causes npm to use
gitignore, preventing this problem in the future. caused npm to ignore
the gitignore, causing it to publish the .env-file which was gitignored
but not npmignored, leaking secrets. Removing this file causes npm to
use gitignore, preventing this problem in the future.

* 0.5.44